deleted dots2.py, migrated refactored code into dots.py
